+++ changelog.md
@@ -1,3 +1,4 @@
 # Unreleased
 
+* Add special filename pattern `*.ext` that renames only the extension.
 * Add special filename pattern `^^^name` that substitutes the third directory segment from the right with `name`.